hackthon/requirements.txt
# Web Framework
Flask>=2.0.0
flask-sqlalchemy>=3.0.0

# Database
SQLAlchemy>=1.4.0

# API and Model
groq>=0.3.0
transformers>=4.30.0
torch>=2.0.0

# PDF Generation
reportlab>=3.6.0
fpdf>=1.7.2

# Environment and System
python-dotenv>=0.19.0
psutil>=5.8.0

# Date and Time Handling
python-dateutil>=2.8.2

# HTTP Requests
requests>=2.28.0

# File Operations
Werkzeug>=2.0.0

# Database Migration
Flask-Migrate>=4.0.0

# Form Handling
WTForms>=3.0.0